% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Time-Concentration data for two formulas}
\format{A data frame with 24 rows and 8 variables including tablet, group and measurements at six points in time.}
\usage{
data(example_data)
}
\description{
(Artificial) Concentration data for 2 formulas, each including 12 tablets.
Concentrations are measured at six points in time.
}
\keyword{datasets}
